marketing/pyproject.toml
[project]
name = "marketing"
version = "0.1.0"
description = "marketing using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.150.0,<1.0.0",
    "python-pptx>=1.0.2",
]

[project.scripts]
marketing = "marketing.main:run"
run_crew = "marketing.main:run"
train = "marketing.main:train"
replay = "marketing.main:replay"
test = "marketing.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
